backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pymysql==1.0.3
python-dotenv==0.19.0
cryptography==41.0.1